Add files via upload
ProductStoreApp

This is a C# console application that allows a user to select and view properties of products stored in the ProductStore. The user is prompted to enter a comma-separated list of properties they want to view, and the program displays the selected properties for each product in the ProductStore.

How to Run
To run this program, you need to have the .NET Core runtime installed on your system. You can download it from the official website: https://dotnet.microsoft.com/download.

Once you have .NET Core installed, follow these steps to run the program:

Clone or download this repository to your local machine.
Open a command prompt or terminal window and navigate to the directory where the program files are stored.
Type dotnet run and press Enter to run the program.
Follow the prompts to select the properties you want to view.
Usage
When you run the program, you will be prompted to enter a comma-separated list of properties you want to view. If you want to view all properties, simply press Enter without entering any values.

The program will then display the selected properties for each product in the ProductStore.

License
This program is licensed under the MIT License. See the LICENSE file for more information.